+++ tests/testthat/test-codes.R
@@ -1,10 +1,12 @@
-test_that("sumario_xml works", {
-    s <- sumario_xml(format(as.Date("2009/01/01", "%Y/%m/%d"), "%Y%m%d"))
+test_that("sumario_nbo works", {
+    s <- sumario_nbo(format(as.Date("2009/01/01", "%Y/%m/%d"), "%Y%m%d"))
     expect_equal(s, "BOE-S-20090101")
 })
 
-test_that("sumario_nbo works", {
-    s <- sumario_nbo(2019, 242)
+test_that("sumario_cve works", {
+    s <- sumario_cve(2019, 242)
+    expect_equal(s, "BOE-S-2019-242")
+    s <- sumario_cve("2019", "242")
     expect_equal(s, "BOE-S-2019-242")
 })
 
@@ -16,13 +18,13 @@ test_that("elemento works", {
 })
 
 
-test_that("disposicion works", {
-    s <- disposicion(year = 242, number = 242)
+test_that("disposicion_cve works", {
+    s <- disposicion_cve(year = 242, number = 242)
     expect_equal(s, "BOE-A-242-242")
 })
 
-test_that("anuncio works", {
-    s <- anuncio(year = 242, number = 242)
+test_that("anuncio_cve works", {
+    s <- anuncio_cve(year = 242, number = 242)
     expect_equal(s, "BOE-B-242-242")
 })
 
@@ -42,10 +44,10 @@ test_that("PDF code is valid", {
 })
 
 
-test_that("sumario_nbo works", {
-  expect_equal(sumario_nbo(2019, 242), "BOE-S-2019-242")
+test_that("sumario_cve works", {
+  expect_equal(sumario_cve(2019, 242), "BOE-S-2019-242")
   # There shouldn't be summaries/sumarios more than one each day.
-  expect_error(sumario_nbo(2019, 2424))
-  expect_error(sumario_nbo(2019, -1), "numeric format")
-  expect_equal(sumario_nbo(1999, 1), "BOE-S-1999-1")
+  expect_error(sumario_cve(2019, 2424))
+  expect_error(sumario_cve(2019, -1), "numeric format")
+  expect_equal(sumario_cve(1999, 1), "BOE-S-1999-1")
 })